Disentangle some parts between flit_core and flit
