+++ flit/buildapi.py
@@ -1,0 +1,1 @@
+from flit_core.buildapi import *
