+++ flit/common.py
@@ -1,0 +1,1 @@
+from flit_core.common import *
