+++ flit/inifile.py
@@ -1,0 +1,2 @@
+from flit_core.inifile import *
+from flit_core.inifile import _prep_metadata
