+++ flit_core/flit_core/wheel.py
@@ -200,26 +200,3 @@ def make_wheel_in(ini_path, wheel_directory):
 
     log.info("Built wheel: %s", wheel_path)
     return SimpleNamespace(builder=wb, file=wheel_path)
-
-def wheel_main(ini_path, upload=False, verify_metadata=False, repo='pypi'):
-    """Build a wheel in the dist/ directory, and optionally upload it.
-    """
-    dist_dir = ini_path.parent / 'dist'
-    try:
-        dist_dir.mkdir()
-    except FileExistsError:
-        pass
-
-    wheel_info = make_wheel_in(ini_path, dist_dir)
-
-    if verify_metadata:
-        from .upload import verify
-        log.warning("'flit wheel --verify-metadata' is deprecated.")
-        verify(wheel_info.builder.metadata, repo)
-
-    if upload:
-        from .upload import do_upload
-        log.warning("'flit wheel --upload' is deprecated; use 'flit publish' instead.")
-        do_upload(wheel_info.file, wheel_info.builder.metadata, repo)
-
-    return wheel_info
